Bugfix
Don't clean up the dock if it does not exist.

+++ src/window.cpp
@@ -257,7 +257,9 @@ bool sysmenu::on_sort(Gtk::FlowBoxChild* a, Gtk::FlowBoxChild* b) {
 void sysmenu::app_info_changed(GAppInfoMonitor* gappinfomonitor) {
 	app_list = Gio::AppInfo::get_all();
 	flowbox_itembox.remove_all();
-	sysmenu_dock->remove_all();
+
+	if (config_main.dock_items != "")
+		sysmenu_dock->remove_all();
 
 	// Load applications
 	for (auto app : app_list)
